+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/datamigration/caseflag/CaseFlag.java
@@ -1,6 +1,5 @@
-package uk.gov.hmcts.reform.prl.models.dto.caseflag;
+package uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,7 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
-public class Flag {
-    @JsonProperty("FlagDetails")
-    private List<FlagDetail> flagDetails;
+public class CaseFlag {
+    private List<Flag> flags;
 }
