+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/datamigration/caseflag/Flag.java
@@ -1,5 +1,6 @@
-package uk.gov.hmcts.reform.prl.models.dto.caseflag;
+package uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,6 +14,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
-public class CaseFlag {
-    private List<Flag> flags;
+public class Flag {
+    @JsonProperty("FlagDetails")
+    private List<FlagDetail> flagDetails;
 }
