+++ src/test/java/uk/gov/hmcts/reform/prl/services/RefDataUserServiceTest.java
@@ -16,9 +16,9 @@ import uk.gov.hmcts.reform.prl.clients.JudicialUserDetailsApi;
 import uk.gov.hmcts.reform.prl.clients.StaffResponseDetailsApi;
 import uk.gov.hmcts.reform.prl.config.launchdarkly.LaunchDarklyClient;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicListElement;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.CaseFlag;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.Flag;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.FlagDetail;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.CaseFlag;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.Flag;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.FlagDetail;
 import uk.gov.hmcts.reform.prl.models.dto.hearingdetails.CategorySubValues;
 import uk.gov.hmcts.reform.prl.models.dto.hearingdetails.CategoryValues;
 import uk.gov.hmcts.reform.prl.models.dto.hearingdetails.CommonDataResponse;
