+++ src/test/java/uk/gov/hmcts/reform/prl/services/caseflags/CaseFlagMigrationServiceTest.java
@@ -13,9 +13,9 @@ import uk.gov.hmcts.reform.prl.models.caseflags.AllPartyFlags;
 import uk.gov.hmcts.reform.prl.models.caseflags.Flags;
 import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
 import uk.gov.hmcts.reform.prl.models.complextypes.citizen.User;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.CaseFlag;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.Flag;
-import uk.gov.hmcts.reform.prl.models.dto.caseflag.FlagDetail;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.CaseFlag;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.Flag;
+import uk.gov.hmcts.reform.prl.models.dto.datamigration.caseflag.FlagDetail;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.services.RefDataUserService;
 import uk.gov.hmcts.reform.prl.services.SystemUserService;
